Moved all enumerations to a dedicated namespace (#304)
