+++ LeoCorpLibrary/UnitsConversions.cs
@@ -22,7 +22,7 @@ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE. 
 */
 
-using LeoCorpLibrary.Extensions;
+using LeoCorpLibrary.Enums;
 using System;
 
 namespace LeoCorpLibrary
